requirements.txt
python-dotenv~=1.0.1
discord-py-interactions
urllib3
requests~=2.31.0
typing_extensions~=4.9.0